+++ android/app/build.gradle.kts
@@ -24,8 +24,6 @@ android {
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_11
         targetCompatibility = JavaVersion.VERSION_11
-
-        isCoreLibraryDesugaringEnabled = true
     }
 
     kotlinOptions {
@@ -48,7 +46,4 @@ flutter {
 dependencies {
     implementation(platform("com.google.firebase:firebase-bom:33.13.0"))
     implementation("com.google.firebase:firebase-analytics")
-    implementation("com.google.firebase:firebase-messaging")
-
-    coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:2.0.4")
 }
