+++ firebase.json
@@ -1,38 +1,1 @@
-{
-  "flutter": {
-    "platforms": {
-      "android": {
-        "default": {
-          "projectId": "pawpal-application-project",
-          "appId": "1:601569258432:android:41bc88a007c0eb6a9ef625",
-          "fileOutput": "android/app/google-services.json"
-        }
-      },
-      "dart": {
-        "lib/firebase_options.dart": {
-          "projectId": "pawpal-application-project",
-          "configurations": {
-            "android": "1:601569258432:android:41bc88a007c0eb6a9ef625",
-            "ios": "1:601569258432:ios:8911d14f66414eb09ef625"
-          }
-        }
-      }
-    }
-  },
-  "functions": [
-    {
-      "source": "functions",
-      "codebase": "default",
-      "ignore": [
-        "node_modules",
-        ".git",
-        "firebase-debug.log",
-        "firebase-debug.*.log",
-        "*.local"
-      ],
-      "predeploy": [
-        "npm --prefix \"$RESOURCE_DIR\" run lint"
-      ]
-    }
-  ]
-}
+{"flutter":{"platforms":{"android":{"default":{"projectId":"pawpal-application-project","appId":"1:601569258432:android:41bc88a007c0eb6a9ef625","fileOutput":"android/app/google-services.json"}},"dart":{"lib/firebase_options.dart":{"projectId":"pawpal-application-project","configurations":{"android":"1:601569258432:android:41bc88a007c0eb6a9ef625","ios":"1:601569258432:ios:8911d14f66414eb09ef625"}}}}}}
