+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
-import 'package:firebase_messaging/firebase_messaging.dart';
 import 'theme/theme.dart';
 import 'routes/routes.dart';
 
@@ -12,53 +11,11 @@ void main() async {
     options: DefaultFirebaseOptions.currentPlatform,
   );
 
-  FirebaseMessaging.onBackgroundMessage(_firebaseMessagingBackgroundHandler);
-
-  await FirebaseMessaging.instance.requestPermission();
-
   runApp(const PawPalApp());
 }
 
-Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
-  await Firebase.initializeApp();
-  print('🔙 Handling a background message: ${message.messageId}');
-}
-
-class PawPalApp extends StatefulWidget {
+class PawPalApp extends StatelessWidget {
   const PawPalApp({super.key});
-
-  @override
-  State<PawPalApp> createState() => _PawPalAppState();
-}
-
-class _PawPalAppState extends State<PawPalApp> {
-  @override
-  void initState() {
-    super.initState();
-    _setupFirebaseMessagingHandlers();
-  }
-
-  void _setupFirebaseMessagingHandlers() {
-    FirebaseMessaging.onMessage.listen((RemoteMessage message) {
-      print('📩 Foreground message received:');
-      print('Title: ${message.notification?.title}');
-      print('Body: ${message.notification?.body}');
-    });
-
-    FirebaseMessaging.onMessageOpenedApp.listen((RemoteMessage message) {
-      print('📬 Notification caused app to open.');
-      print('Data: ${message.data}');
-      // Handle navigation based on message data if needed
-    });
-
-    FirebaseMessaging.instance.getInitialMessage().then((message) {
-      if (message != null) {
-        print('📨 App opened from terminated state via notification.');
-        print('Data: ${message.data}');
-        // Handle navigation based on message data if needed
-      }
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
