+++ lib/routes/routes.dart
@@ -3,36 +3,16 @@
 import 'package:flutter/material.dart';
 import '../screens/login_screen.dart';
 import '../screens/register_screen.dart';
+import '../screens/discovery_screen.dart';
 import '../screens/forgot_password_screen.dart';
 import '../screens/email_verification_screen.dart';
-import '../screens/main_screen.dart';
-import '../screens/discovery_screen.dart';
-import '../screens/chat_screen.dart';
-import '../screens/user_profile_screen.dart';
-import '../screens/edit_user_profile_screen.dart';
-import '../screens/user_settings_screen.dart';
-import '../screens/user_faq_screen.dart';
-import '../screens/user_privacy_policy_screen.dart';
-import '../screens/user_terms_screen.dart';
-import '../screens/about_the_app_screen.dart';
-import '../screens/user_notifications_screen.dart';
 
 class AppRoutes {
   static const String login = '/login';
   static const String register = '/register';
+  static const String discovery = '/discovery';
   static const String forgotPassword = '/forgot-password';
   static const String emailVerification = '/email-verification';
-  static const String mainScreen = '/main-screen';
-  static const String discovery = '/discovery';
-  static const String chat = '/chat';
-  static const String userProfile = '/user-profile';
-  static const String editUserProfile = '/edit-user-profile';
-  static const String userSettings = '/user-settings';
-  static const String userFaq = '/user-faq';
-  static const String userPrivacyPolicy = '/user-privacy-policy';
-  static const String userTerms = '/user-terms';
-  static const String aboutTheApp = '/about-the-app';
-  static const String userNotifications = '/user-notifications';
 
   static Route<dynamic> generateRoute(RouteSettings settings) {
     switch (settings.name) {
@@ -40,32 +20,12 @@ class AppRoutes {
         return MaterialPageRoute(builder: (_) => const LoginScreen());
       case register:
         return MaterialPageRoute(builder: (_) => const RegisterScreen());
+      case discovery:
+        return MaterialPageRoute(builder: (_) => const DiscoveryScreen());
       case forgotPassword:
         return MaterialPageRoute(builder: (_) => const ForgotPasswordScreen());
       case emailVerification:
         return MaterialPageRoute(builder: (_) => const EmailVerificationScreen());
-      case mainScreen:
-        return MaterialPageRoute(builder: (_) => const MainScreen());
-      case discovery:
-        return MaterialPageRoute(builder: (_) => const DiscoveryScreen());
-      case chat:
-        return MaterialPageRoute(builder: (_) => const ChatScreen());
-      case userProfile:
-        return MaterialPageRoute(builder: (_) => const UserProfileScreen());
-      case editUserProfile:
-        return MaterialPageRoute(builder: (_) => const EditUserProfileScreen());
-      case userSettings:
-        return MaterialPageRoute(builder: (_) => const UserSettingsScreen());
-      case userFaq:
-        return MaterialPageRoute(builder: (_) => const UserFaqScreen());
-      case userPrivacyPolicy:
-        return MaterialPageRoute(builder: (_) => const UserPrivacyPolicyScreen());
-      case userTerms:
-        return MaterialPageRoute(builder: (_) => const UserTermsScreen());
-      case aboutTheApp:
-        return MaterialPageRoute(builder: (_) => const AboutTheAppScreen());
-      case userNotifications:
-        return MaterialPageRoute(builder: (_) => const UserNotificationsScreen());
 
       default:
         return MaterialPageRoute(
