+++ lib/screens/user_profile.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_database/firebase_database.dart';
+import '../firebase_options.dart';
+import '../theme/theme.dart';
+import '../routes/routes.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+// UserProfileScreen
+// This screen displays the user's profile information and allows them to edit their profile.
