+++ pubspec.yaml
@@ -39,9 +39,6 @@ dependencies:
   firebase_auth: ^5.5.3
   firebase_database: ^11.3.5
   shared_preferences: ^2.5.3
-  image_picker: ^1.1.2
-  http: ^1.4.0
-  firebase_messaging: ^15.2.5
 
 dev_dependencies:
   flutter_test:
